:truck: refactor: entity폴더 domain폴더로 이름 변경
